refactor: Add link to category card
