+++ src/components/Molucules/CategoryCard/CategoryCard.type.ts
